Move Playwright execution into its node handler
The CI file line-limit check failed: workflow_executor.py reached 10005
lines once this branch merged with main. The Playwright path was the
largest node-specific block still living there, which AGENTS.md already
asks to keep under node_execution/nodes/.

Script assembly, the scrubbed subprocess environment, auth-state
resolution, expression substitution and the sandbox/subprocess run all
move to playwright_node.py as module functions taking the executor. No
behavior change; the executor is back to 9604 lines.
