+++ backend/tests/test_playwright_global_expressions.py
@@ -10,8 +10,12 @@ placeholder instead of navigating to the configured URL.
 import re
 import unittest
 
+from app.services.node_execution.nodes.playwright_node import (
+    _build_playwright_script,
+    _playwright_subprocess_inputs,
+)
 from app.services.playwright_code_generator import generate_playwright_code
-from app.services.workflow_executor import WorkflowExecutor, _build_playwright_script
+from app.services.workflow_executor import WorkflowExecutor
 
 
 def _executor(global_variables_context: dict[str, object] | None = None) -> WorkflowExecutor:
@@ -32,7 +36,7 @@ def _resolved_goto_urls(steps: list[dict], inputs: dict) -> list[object]:
 class PlaywrightGlobalNamespaceTests(unittest.TestCase):
     def test_global_namespace_is_injected_into_runner_inputs(self) -> None:
         executor = _executor({"secondPage": "https://heym.run/pricing"})
-        runner_inputs = executor._playwright_subprocess_inputs({})
+        runner_inputs = _playwright_subprocess_inputs(executor, {})
         self.assertEqual(runner_inputs["global"]["secondPage"], "https://heym.run/pricing")
 
     def test_variable_node_values_are_merged_into_global(self) -> None:
@@ -40,13 +44,13 @@ class PlaywrightGlobalNamespaceTests(unittest.TestCase):
         executor = _executor({"secondPage": "https://heym.run/pricing", "keep": "kept"})
         executor.vars = {"secondPage": "https://heym.run/solutions"}
         executor._mark_vars_context_dirty()
-        runner_inputs = executor._playwright_subprocess_inputs({})
+        runner_inputs = _playwright_subprocess_inputs(executor, {})
         self.assertEqual(runner_inputs["global"]["secondPage"], "https://heym.run/solutions")
         self.assertEqual(runner_inputs["global"]["keep"], "kept")
 
     def test_navigate_step_uses_global_url_not_placeholder(self) -> None:
         executor = _executor({"secondPage": "https://heym.run/pricing"})
-        runner_inputs = executor._playwright_subprocess_inputs({})
+        runner_inputs = _playwright_subprocess_inputs(executor, {})
         steps = [
             {"action": "navigate", "url": "https://heym.run/solutions"},
             {"action": "navigate", "url": "$global.secondPage"},
@@ -58,8 +62,8 @@ class PlaywrightGlobalNamespaceTests(unittest.TestCase):
 
     def test_upstream_node_named_global_is_not_clobbered(self) -> None:
         executor = _executor({"secondPage": "https://heym.run/pricing"})
-        runner_inputs = executor._playwright_subprocess_inputs(
-            {"global": {"ownField": "from-upstream"}}
+        runner_inputs = _playwright_subprocess_inputs(
+            executor, {"global": {"ownField": "from-upstream"}}
         )
         self.assertEqual(runner_inputs["global"]["ownField"], "from-upstream")
         self.assertEqual(runner_inputs["global"]["secondPage"], "https://heym.run/pricing")
@@ -68,7 +72,7 @@ class PlaywrightGlobalNamespaceTests(unittest.TestCase):
         executor = _executor()
         executor.vars = {"searchUrl": "https://heym.run/docs"}
         executor._mark_vars_context_dirty()
-        runner_inputs = executor._playwright_subprocess_inputs({})
+        runner_inputs = _playwright_subprocess_inputs(executor, {})
         steps = [{"action": "navigate", "url": "$vars.searchUrl"}]
         self.assertEqual(_resolved_goto_urls(steps, runner_inputs), ["https://heym.run/docs"])
